+++ regionmask/core/mask.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from .utils import _wrapAngle
+from .utils import _wrapAngle, _is_180
 
 
 def _maybe_import_xarray():
@@ -23,7 +23,7 @@ def _mask(
     lon_name="lon",
     lat_name="lat",
     xarray=None,
-    wrap_lon=False,
+    wrap_lon=None,
 ):
     """
     create a grid as mask of a set of regions for given lat/ lon grid
@@ -79,6 +79,13 @@ def _mask(
 
     lon = np.array(lon)
     lat = np.array(lat)
+
+    # automatically detect whether wrapping is necessary
+    if wrap_lon is None:
+        regions_is_180 = self.lon_180
+        grid_is_180 = _is_180(lon.min(), lon.max())
+
+        wrap_lon = not regions_is_180 == grid_is_180
 
     if wrap_lon:
         lon_old = lon.copy()